Convert status embed controller to request spec (#32448)
